ENGINE: add `engine_session` to make type changing directly in cheater mode possible
